I messed up. Again
Turns out the 1D Heat equation stationnary was working, until I changed something that broke the training function.

What broke everything was :
- adding .detach() with the loss
- adding a * (x_max - x_min) + x_min on the inputs vector (line 8)

It seems to screw up the requires_grad value of the tensor, and it is then impossible to do a .backwards() on the loss.

I fixed it by changing the requires_grad_ value to True **AFTER** I created the inputs tensor.

may my soul be blessed with this founding