Currently questioning my carreer choices
I have paused fixing the 1D heat eq with time, and focused on the stationary one.

For (8,20) and (2,5) architecture, the predictions seemed fine.
For the (10, 100), we got some strange predictions from time to time, and there are several leads to explain why (see report for more info).

The MSE doesn't seem to work, as I had a similar one for two predictions who were clearly not as accurate as the other, so this metric might be broken.
I will assess the predictions visually for now.

Plotting the last prediction didn't seem necessary, as the checkpoint is supposed to be the best prediction. I will remove it from now.

I think the network is just too big for what I am trying to do. Please don't fatshame him.