Still working on the time prediction
Reworked the boundary_conditions() because the sizes of the tensors seemed to make errors.
Turns out I had to rework the other functions used in the loss_composite().

Still making changes to these functions so that everything runs fine with time.

Added a "saved_models" directory in each of the different equations directory, so we can copy the "good" models and not overwrite them each time we launch a training.

Also removed some unnecessary lines (like WandB commented logs and u_predict stuff, that we don't display anymore).